Added coins, shrank the bolder, and forward pushes
The rock now has a initial force value slider which describes the amount
of force the rock starts with. There is also a forward force and number
of forward pushes which allows the player to add a certain amount of
forward force to the rock a certain amount of times. These are public so
the upgrade system can eventually allow them to change throughout the
game. Coins are also in the scenes now and increase the player's money
by 5 each.