Add documentation, new money script, and more
Currently, we do not have a money UI counter script. Add one so we can
have a constant money counter going forward. Also, this will allow us to
not have to manually update the money counter after every upgrade
purchase. Also, cleaned some of the useless comments out of the scripts
and made it so the money counter on the upgrade screen subtracts after
each purchase.