Added rock size upgrade, minor UI's, coin fixes
The rock's size can now be upgraded within the upgrade system with its
max size being close to a scale of 3. I've also added a in game UI so
the player can see how many forward and upward pushes they have. Finally
I fixed the coins so that they had a consistent size and removed some
extreme peaks/dips in the new landmasses.