Seperated Pushes and Force
Seperated pushes and force into two seperate upgrades for both upward
and forward pushes and added both to the UI. This means that the force
applied with a forward push is no longer tied to the amount of forward
pushes you have, and the same for upward pushes. Also minor landmass
fixes/smoothing. I also moved alot of control from the movement script
to the restart/game controller script such as the inital number of
pushes and initial amount of force applied by each type of force.