Finished moving upgrade control
Removed commented out lines that were for testing moving all upgrade
control to the restart game script. The movement script now just simply
controls adding the force to the rock when pushes are used, saving the
upgrades and calculating the values to use for upgrades is now contained
in the restart game script.