Changed coins and rock physics
Changed coins to use the new sprite, this might have changed some of
their placements slightly since I had to resize the coins to match up
with the new sprite. Also the rock's collider should now prevent it from
breaking the game without the cooldown between pushes, but I'm leaving
the cooldown in as a gameplay element for now.